guacatea_market/requirements.txt
alembic==1.7.5
aniso8601==9.0.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==2.0.2
Flask-Alembic==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.24.1
Pillow==9.0.1
psycopg2-binary==2.9.1
pycparser==2.21
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.26
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.7.0